max1720x_battery: correct EEPROM history maxdischgcurr unit
The calculation for hist.maxtemp and hist.mintemp were using
unsigned values, change to signed values to ensure correct
values.

Bug: 290990869

+++ max1720x_battery.c
@@ -5436,17 +5436,17 @@ static int max17x0x_collect_history_data(void *buff, size_t size,
 		return ret;
 
 	/* Convert LSB from 1degC to 3degC, store values from 25degC min */
-	hist.maxtemp = (REG_HALF_HIGH(data) - 25) / 3;
+	hist.maxtemp = ((s8)REG_HALF_HIGH(data) - 25) / 3;
 	/* Convert LSB from 1degC to 3degC, store values from -20degC min */
-	hist.mintemp = (REG_HALF_LOW(data) + 20) / 3;
+	hist.mintemp = ((s8)REG_HALF_LOW(data) + 20) / 3;
 
 	ret = REGMAP_READ(&chip->regmap, MAX1720X_MAXMINCURR, &data);
 	if (ret)
 		return ret;
 
 	/* Convert LSB from 0.08A to 0.5A */
-	hist.maxchgcurr = REG_HALF_HIGH(data) * 8 / 50;
-	hist.maxdischgcurr = REG_HALF_LOW(data) * 8 / 50;
+	hist.maxchgcurr = (s8)REG_HALF_HIGH(data) * 8 / 50;
+	hist.maxdischgcurr = (s8)REG_HALF_LOW(data) * 8 / 50;
 
 	memcpy(buff, &hist, sizeof(hist));
 	return (size_t)sizeof(hist);
